+++ CHANGES.md
@@ -317,6 +317,10 @@
 
     - Added `doLower` and `doRaise`
 
+    - Added `shiftToSame` and `clientLeader` which allow a hook to be created
+      that shifts a window to the workspace of other windows of the application
+      (using either the `WM_CLIENT_LEADER` or `_NET_WM_PID` property).
+
   * `XMonad.Util.EZConfig`
 
     - Added support for XF86Bluetooth.